move tests for reporting
